pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bark"
version = "0.1.0"
description = "A static site generator"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Ashley Cottrell" },
]
keywords = ["static-site-generator", "markdown", "cli", "collections"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
]
dependencies = [
    "typer>=0.15.0,<1.0",
    "jinja2>=3.1,<4.0",
    "markdown>=3.7,<4.0",
    "pyyaml>=6.0,<7.0",
    "pygments>=2.18,<3.0",
    "watchdog>=6.0,<7.0",
    "rich>=13.0,<14.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-cov>=6.0,<7.0",
]

[project.scripts]
bark = "bark.cli:app"

[project.urls]
Homepage = "https://github.com/cottrellashley/bark"
Repository = "https://github.com/cottrellashley/bark"
Issues = "https://github.com/cottrellashley/bark/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/bark"]

[tool.deptry.package_module_name_map]
typer = "typer"
markdown = "markdown"
pyyaml = "yaml"
